+++ tests/test_next.rs
@@ -1,5 +1,3 @@
-#![feature(generic_associated_types)]
-
 use unimock::*;
 
 use async_trait::async_trait;